requirements.txt
PyPDF2
pypdf
pdfminer.six
pdfplumber
spacy
nltk
pandas
openpyxl
tqdm
openai
python-dotenv
pyyaml
json_repair
gradio
paddlex[ocr]>=3.4.1
